+++ crates/integrations/connector-integration/src/payout_connectors.rs
@@ -10,6 +10,9 @@ pub use self::paypal::PaypalPayouts;
 pub mod deutschebank;
 pub use self::deutschebank::DeutschebankPayouts;
 
+pub mod truelayer;
+pub use self::truelayer::TruelayerPayouts;
+
 pub mod worldpayxml;
 pub use self::worldpayxml::WorldpayxmlPayouts;
 
@@ -18,3 +21,6 @@ pub use self::cybersource::CybersourcePayouts;
 
 pub mod santander;
 pub use self::santander::SantanderPayouts;
+
+pub mod trustly;
+pub use self::trustly::TrustlyPayouts;